Fix exception when directory not created yet.

+++ src/XRoadLib.Tools/Program.fs
@@ -15,7 +15,7 @@ let main args =
             async {
                 let options = optionsAccessor.Value
                 let! schemaDoc = loadSchema options
-                options.OutputPath.Delete(true)
+                if options.OutputPath.Exists then options.OutputPath.Delete(true)
                 genServiceCode options schemaDoc
                 return 0
             } |> Async.StartAsTask)
